add bar_chart by charts_flutter
