+++ lib/bar_chart_demo.dart
@@ -1,0 +1,73 @@
+import 'package:flutter/material.dart';
+import 'package:charts_flutter/flutter.dart' as charts;
+import 'package:handrange/bar_chart_graph.dart';
+import 'package:handrange/bar_chart_model.dart';
+
+
+class BarChartDemo extends StatefulWidget {
+  @override
+  _BarChartDemoState createState() => _BarChartDemoState();
+}
+
+class _BarChartDemoState extends State<BarChartDemo> {
+
+  final List<BarChartModel> data = [
+    BarChartModel(
+      year: "2016",
+      financial: 100,
+      color: charts.ColorUtil.fromDartColor(Colors.green),
+    ),
+
+    BarChartModel(
+      year: "2017",
+      financial: 450,
+      color: charts.ColorUtil.fromDartColor(Colors.yellow),
+    ),
+    BarChartModel(
+      year: "2018",
+      financial: 630,
+      color: charts.ColorUtil.fromDartColor(Colors.lightBlueAccent),
+    ),
+    BarChartModel(
+      year: "2019",
+      financial: 600,
+      color: charts.ColorUtil.fromDartColor(Colors.pink),
+    ),
+    BarChartModel(
+      year: "2020",
+      financial: 400,
+      color: charts.ColorUtil.fromDartColor(Colors.purple),
+    ),
+    BarChartModel(
+      year: "2021",
+      financial: 100,
+      color: charts.ColorUtil.fromDartColor(Colors.green),
+    ),
+  ];
+
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text("Flutter Animated Bar Chart Demo"),
+        automaticallyImplyLeading: false,
+        centerTitle: true,
+      ),
+      body: Center(
+        child: Container(
+          child: ListView(
+            scrollDirection: Axis.horizontal,
+            shrinkWrap: true,
+            children: [
+              BarChartGraph(
+                data: data,
+              ),
+            ],
+          ),
+
+        ),
+      ),
+    );
+  }
+}
